+++ NEWS.md
@@ -1,3 +1,6 @@
 # ohsome 0.2.0
 
 * Added a `NEWS.md` file to track changes to the package.
+* Changed behaviour of set_properties(): Removes properties parameter from query 
+body by default, accepts "tags" and/or "metadata" as properties argument 
+(multiple values provided as comma-separated character or character vector).